+++ protected/humhub/docs/CHANGELOG.md
@@ -5,6 +5,8 @@ HumHub Change Log
 --------------------------
 
 - Fix: img overflow in markdown view
+- Enh: Added console command to list and change themes
+
 
 1.3.2  (September 4, 2018)
 --------------------------